requirements.txt
Django==4.2.7
djangorestframework==3.14.0
langgraph==0.0.20
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1 